Feature/ant route search algorithm

Fix #22

See merge request !17

+++ PathFinder.Trips.WebApi/Models/TripReadModel.cs
@@ -24,6 +24,6 @@ namespace PathFinder.Trips.WebApi.Models
     public class Route
     {
         public IList<int> Sequence { get; set; }
-        public int Distanse { get; set; }
+        public double Distanse { get; set; }
     }
 }
